# src/authserver/web/openapi.yaml
openapi: 3.0.3
info:
  title: Goiabada API
  description: |
    REST API for Goiabada authentication and authorization server.

    ## Authentication

    Every operation under `/api/v1` requires a valid JWT access token in the
    `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    `getClientLogoImage` is the exception and is marked `security: []`: it is public.

    ## API Groups

    - **Admin API** (`/api/v1/admin/*`): administrative control. Each operation accepts any
      one of a small set of scopes rather than a single one: reads accept
      `authserver:admin-read` alongside the domain's own scope, and writes accept the
      domain scope (`authserver:manage-users`, `authserver:manage-clients` or
      `authserver:manage-settings`). `authserver:manage` is accepted everywhere.
      A token from the client credentials flow is accepted here.
    - **Account API** (`/api/v1/account/*`): Self-service account management. Requires
      `authserver:manage-account` scope, **and** a token issued for a user: these operations
      resolve the acting user from `sub`, so a client credentials token is refused with 403
      and `error_code` `USER_CONTEXT_REQUIRED`.

    ## Caching

    Every response from an operation in either API group carries these headers, error
    responses and the token and scope refusals included:
    ```
    Cache-Control: no-store
    Pragma: no-cache
    ```

...

...
    none of them is described by this document.

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/admin/group-attributes:
    post:
      tags: [Group Attributes]
      summary: Create group attribute
      description: Create a new custom attribute for a group
      operationId: createGroupAttribute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupAttributeRequest'
      responses:
        '201':
          description: Attribute created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGroupAttributeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/group-attributes/{id}:
    get:
      tags: [Group Attributes]
      summary: Get group attribute
      description: Get a specific group attribute by ID
      operationId: getGroupAttribute
      parameters:
        - name: id
          in: path
          required: true
          description: Attribute ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Attribute details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetGroupAttributeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Group Attributes]
      summary: Update group attribute
      description: Update a group attribute
      operationId: updateGroupAttribute
      parameters:
        - name: id
          in: path
          required: true
          description: Attribute ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupAttributeRequest'
      responses:
        '200':
          description: Attribute updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateGroupAttributeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Group Attributes]
      summary: Delete group attribute
      description: Delete a group attribute
      operationId: deleteGroupAttribute
      parameters:
        - name: id
          in: path
          required: true
          description: Attribute ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Attribute deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # =============================================================================
  # ADMIN API - RESOURCES
  # =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token.

        Admin API: any one of `authserver:admin-read`, the domain scope for the operation
        (`authserver:manage-users`, `authserver:manage-clients`, `authserver:manage-settings`)
        or `authserver:manage`. A client credentials token is accepted.

        Account API: scope `authserver:manage-account`, on a token issued for a user. A client
        credentials token carries no `auth_time` claim and is refused with 403
        `USER_CONTEXT_REQUIRED`.

        Browser sessions: scope `authserver:browser-sessions`, and only that scope. It is
        deliberately not one of the manage-* scopes, so holding the admin console's client secret
        is not a way to drive the admin API with no user present.
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: |
        The token is valid but does not authorise this request. Three checks
        answer here, and a caller distinguishes them by error_code.

        INSUFFICIENT_SCOPE comes from the route's authorisation middleware,
        before the handler runs: an Admin operation needs one of the scopes
        its route names, and an Account operation needs
        `authserver:manage-account`.

        USER_CONTEXT_REQUIRED comes from the same middleware on the Account
        operations, which also require the token to have been issued for a
        user: one obtained through the client credentials grant carries no
        auth_time claim and is refused here.

        FORBIDDEN comes from the handler, on the account endpoints that
        resolve a resource by id and then check it belongs to the caller, so
        a valid token for the wrong user reaches this rather than a 404.

        Every operation this document secures declares a 403. The public
        client logo endpoint, which carries `security: []` and authorises
        nothing, is the one that does not.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: |
        The request failed inside the server, most often a storage error.
        error_code is INTERNAL_SERVER_ERROR, except on key rotation, which also
        answers KEY_SET_INCOMPLETE. The description carries no detail about the
        failure, only the request id an operator can find the log line by.
        Retrying is reasonable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Admin/account API error envelope. Routing by HTTP status code:
        4xx is a client-correctable failure (validation, not found,
        permission denied, rate limited); 5xx is a server-side problem.
        The body provides a stable error_code, optional error_args, and
        an English error_description for non-localizing consumers.
        Protocol endpoints (/auth/token, /auth/authorize, /connect/register,
        /userinfo) use their RFC-defined error envelopes instead of this one.
      properties:
        error_code:
          type: string
          description: |
            Stable identifier for the failure. UPPER_SNAKE for legacy codes
            (e.g. VALIDATION_ERROR), dotted lowercase for catalog-keyed
            localized codes (e.g. validator.email.required).
        error_args:
          type: object
          additionalProperties: true
          description: Substitution values for the localized message template.
        error_description:
          type: string
          description: Rendered English message. Stable for non-localizing consumers.
      required:
        - error_description
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # =========================================================================
    # User Schemas
    # =========================================================================
...
    GroupAttributeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        key:
          type: string
        value:
          type: string
        includeInIdToken:
          type: boolean
        includeInAccessToken:
          type: boolean
        groupId:
          type: integer
          format: int64
      required:
        - id
        - createdAt
        - updatedAt
        - key
        - value
        - includeInIdToken
        - includeInAccessToken
        - groupId
...
    GetGroupAttributeResponse:
      type: object
      properties:
        attribute:
          $ref: '#/components/schemas/GroupAttributeResponse'
      required:
        - attribute
    CreateGroupAttributeRequest:
      type: object
      required:
        - key
        - value
        - groupId
      properties:
        key:
          type: string
        value:
          type: string
        includeInIdToken:
          type: boolean
        includeInAccessToken:
          type: boolean
        groupId:
          type: integer
          format: int64
    CreateGroupAttributeResponse:
      type: object
      properties:
        attribute:
          $ref: '#/components/schemas/GroupAttributeResponse'
      required:
        - attribute
    UpdateGroupAttributeRequest:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        includeInIdToken:
          type: boolean
        includeInAccessToken:
          type: boolean
    UpdateGroupAttributeResponse:
      type: object
      properties:
        attribute:
          $ref: '#/components/schemas/GroupAttributeResponse'
      required:
        - attribute

    # =========================================================================
    # Resource & Permission Schemas
    # =========================================================================
...